server/requirements.txt
flask
flask-cors
anthropic
python-dotenv
gunicorn
groq